Allow block-less expectations
This patch allows us to write

expect(3) < 4

instead of

expect true do
  3 < 4
end

and

expect(Pathname('a') / Pathname('b')) == Pathname('a') + Pathname('b')

instead of

expect Pathname('a') + Pathname('b') do
  (Pathname('a') / Pathname('b'))
end

One could already do this with #to.be:

expect 3.to.be < 4

but this is less verbose.

One thing to note is that the parentheses are required (as Ruby doesn’t
have macros):

expect 3 < 4

works fine as long as this holds, but

expect 4 < 4

will generate “Expected false”, which may a bit confusing.