Rework Expectation hierarchy
It makes more sense, I hope, to have Behavior and State as classes and
Expectations as a module.  We also dump the use of accessors, as I feel
that it’s easier to understand if you simply set instance variables
instead.  Perhaps self.message = … still makes sense, as this is a
method that is added by the extend Lookout::Results::…, but it was more
uniform to use instance variables, so let’s give it a try.