Simplify Groups and operations
This changes the groups algorithm implementation to not require storing
all the operations before applying it.  This is done by saving the group
that is about to be yielded and only yielding it if we see another group
or if it was the last one.  This is so that we can fold the last group
appropriately.  The same is done for the first group, but is a lot
easier to test for.

The various operations are also made a bit more intelligent.