Use .=== instead of #is_a? in Equality
The reason for using

        Array === actual

instead of

        actual.is_a? Array

is that actual might not respond to #is_a?, which then raises an error.
The latter reads a lot better, but we don’t want errors to be raised.

We also need to rescue any errors raised from #respond_to? (which the
object might not respond to either).