Remove delegation support
This was a long time coming.  I have come to the conclusion that
delegation isn’t something you should be testing.  Sure, you want to
make sure that you get the correct result, but to whom your delegating
the work to produce said result isn’t something a test should be
testing.