Define Expectations in Expectations#load
The Expectations method is now defined in Expectations#load so that it’s only
available during the loading of the file and, what’s better, the Expectations
object itself is able to take care of the proper evaluation taking place.  We
can now easily have different Expectations objects loaded in the same process,
defining Expectations as they see fit.