Split Stub::Method and Mock::Method into Undefined and Defined
It’s better to have separate types for undefined and defined methods.  This
will enable further clean-ups in the future.  This patch is rather dirty.  The
Mock::Methods.define interface is currently horrible, 1.9 specific, and needs
to be fixed.  First, let’s get rid of Recorders::Reception, so that we won’t
need the set-up block to Mock::Methods.define.  Then we can clean it up
properly.