Add some API documentation
I’m far from happy with YARD.  It’s so verbose and repetitive.  I
originally liked the idea of type annotiations, but I’m beginning to
feel that it interferes with both the description and the code.  The
types of a method’s arguments should be obvious from the known
types (such as the class/module being discussed) and the method’s name.

I’ll have to look into this further.